Replace v1 with v2 components
