+++ packages/shared-ux/chrome/navigation/src/ui/components/navigation_footer.tsx
