+++ packages/shared-ux/chrome/navigation/src/ui/hooks/use_init_navnode.ts
