+++ packages/shared-ux/chrome/navigation/src/ui/hooks/use_register_tree_node.ts
