+++ packages/shared-ux/chrome/navigation/src/ui/nav_tree_presets/analytics.ts
